added litedom.builder

+++ scripts/generateTestHTML.js
@@ -1,10 +1,11 @@
 import { mkdir, writeFile } from "node:fs/promises";
 import { dirname, basename } from "node:path";
 import { entries } from "./entries.js";
+import { existsSync } from "node:fs";
 for (const entry of entries) {
   const path = "test/html/" + entry + ".html";
   await mkdir(dirname(path), { recursive: true });
-  writeFile(
+  if (!existsSync(path)) writeFile(
     path,
     `<!doctype html>
 <script type=module>
