+++ scripts/generateTestHTML.ts
@@ -1,11 +1,12 @@
 import { mkdir, writeFile } from 'node:fs/promises';
 import { dirname, basename } from 'node:path';
 import { entries } from './entries.ts';
+import { existsSync } from 'node:fs';
 
 for (const entry of entries) {
 	const path = 'test/html/' + entry + '.html';
 	await mkdir(dirname(path), { recursive: true });
-	writeFile(path, 
+	if (!existsSync(path)) writeFile(path, 
 `<!doctype html>
 <script type=module>
 	import * as mod from '${'/src/' + entry + '.ts'}';
